task104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 int main(void)
 {
 	int i;
-	unsigned int y, x, z;
+	unsigned long int y, x, z;
 
 	y = 1;
 	x = 1;
 	z = 1;
 	for (i = 1; i <= 98; i++)
 	{
-		printf("%u", z);
+		printf("%lu", z);
 		z = y + x;
 		x = y;
 		y = z;
